+++ cli/src/main/kotlin/sift/cli/TemplateOptions.kt
@@ -4,14 +4,13 @@ import com.github.ajalt.clikt.completion.CompletionCandidates
 import com.github.ajalt.clikt.parameters.groups.OptionGroup
 import com.github.ajalt.clikt.parameters.options.convert
 import com.github.ajalt.clikt.parameters.options.flag
-import com.github.ajalt.clikt.parameters.options.help
 import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.types.path
-import java.nio.file.Path
-import kotlin.io.path.exists
+import org.objectweb.asm.tree.ClassNode
+import sift.core.asm.resolveClassNodes
+import sift.core.template.SystemModelTemplate
 
 class TemplateOptions : OptionGroup(name = "Template options") {
-    val template by option("-t", "--template",
+    val template: SystemModelTemplate? by option("-t", "--template",
             metavar = "TEMPLATE",
             help = "The template producing the system model.",
             completionCandidates = CompletionCandidates.Fixed(templateNames().toSet()))
@@ -26,13 +25,11 @@ class TemplateOptions : OptionGroup(name = "Template options") {
             help = "Lists entity types defined by template.")
         .flag()
 
-    val path: Path? by option("-f", "--class-dir",
-            metavar = "PATH",
-            help = "Path to directory structure containing classes or path to .jar",
-            completionCandidates = CompletionCandidates.Path)
-        .path(mustExist = true)
-        .help("Jar or root directory with classes to analyze.")
-        .convert { p -> p.resolve("target/classes").takeIf(Path::exists) ?: p }
+    val classNodes: List<ClassNode>? by option("-f", "--class-dir", "--classes",
+            metavar = "PATH|URI",
+            help = "Path to directory structure containing classes or path to 'jar' file. " +
+                   "If the path is a URI, it must point to a 'jar' file.")
+        .convert { resolveClassNodes(it) }
 
     val profile: Boolean by option("--profile",
             help = "Print execution times and input/output for the executed template.")